#2 Fixes the minimum spacing thing
It was such an easy fix, I didn't even look it up. It just came up while I was playing around in the editor.
You can set a min-width and min-height in css for certain elements or classes.

Who would have thought.

Also I added some scrolling for the text box element. It's not working very well for now...